Adds useravatar component and fixes router navlink
