+++ src/components/Nav/_Nav.scss
@@ -9,29 +9,36 @@
     display: flex;
     justify-content: space-between;
     align-items: flex-end;
-    height: 100%;
+    height: 60px;
   }
-  ul {
+  .left-side {
     list-style: none;
     margin: 0;
-    li {
-      display: inline-block;
-    }
     a {
+      display: inline-block;
       color: $grey;
       font-size: 16px;
       text-transform: uppercase;
       font-weight: 600;
       padding: 10px 0;
       margin: 0 14px;
-      display: block;
       text-decoration: none;
-    }
-    .selected {
-      a {
+      &.active {
         color: #6E7377;
         border-bottom: solid 4px #A7E1FF;
       }
+    }
+  }
+  .user-link {
+    height: 62px;
+    display: flex;
+    align-items: center;
+    text-decoration: none;
+    .user-username {
+      color: #1F272E;
+      font-weight: 600;
+      text-transform: uppercase;
+      margin-right: 37px;
     }
   }
 }
