+++ src/components/UserAvatar/UserAvatar.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function UserAvatar() {
+  return (
+    <div className="user-avatar"></div>
+  );
+ }
